rename: 재사용성을 위한 orders-payments 하위 컴포넌트의 위치 이동 (#179)
