+++ src/components/orders-payments/index.ts
@@ -2,5 +2,8 @@ export { default as CheckoutHeader } from './CheckoutHeader';
 export { default as DeliveryAddress } from './DeliveryAddress';
 export { default as PaymentMethodSelector } from './PaymentMethodSelector';
 export { default as PaymentSummary } from './PaymentSummary';
-export { default as SectionHeadingTitle } from './SectionHeadingTitle';
 export * from './OrderProductList';
+export * from './OrderHistoryList';
+export { default as ProductName } from './ProductName';
+export { default as ProductAmoutInfo } from './ProductAmoutInfo';
+export { default as ProductOptionInfo } from './ProductOptionInfo';
